+++ server/src/store/value.rs
@@ -2,7 +2,7 @@ use bincode_next::{Encode, Decode};
 
 #[derive(Debug, Clone, PartialEq, Encode, Decode)]
 pub enum Value {
-    Int(i64),
+    Int(isize),
     Float(f64),
     Text(String),
     Blob(Vec<u8>),
@@ -11,7 +11,7 @@ pub enum Value {
 }
 
 impl Value {
-    pub fn int(n: i64) -> Self { Value::Int(n) }
+    pub fn int(n: isize) -> Self { Value::Int(n) }
     pub fn float(n: f64) -> Self { Value::Float(n) }
     pub fn text(s: &str) -> Self { Value::Text(s.to_owned()) }
     pub fn blob(b: Vec<u8>) -> Self { Value::Blob(b) }
@@ -32,6 +32,6 @@ impl Value {
 
 // Implement this later
 pub enum Key {
-    Int(i64),
+    Int(isize),
     Text(String),
 }
